refactor: extract directive navigation from definition.ts
Split 184-line handleDirectiveNavigation function into its own
definition-directives.ts module.

Before: definition.ts = 1,441 lines
After:  definition.ts = 1,255 lines
        definition-directives.ts = 152 lines

Directive handling covers #include, import, inherit, #require.
