+++ packages/pike-lsp-server/src/features/navigation/definition.ts
@@ -15,6 +15,7 @@ import { Logger } from '@pike-lsp/core';
 import { extractExpressionAtPosition } from './expression-utils.js';
 import type { ExpressionInfo, PikeSymbol, InheritanceInfo } from '@pike-lsp/pike-bridge';
 import { readFile } from 'node:fs/promises';
+import { handleDirectiveNavigation } from './definition-directives.js';
 
 const uriDecodeLog = new Logger('Navigation');
 
@@ -768,194 +769,7 @@ async function findSymbolInDirectIncludes(
   return null;
 }
 
-/**
- * Handle navigation for directive lines (#include, import, inherit).
- * Called early to avoid expression extraction mangling include paths like "../foo.h"
- * into dotted expressions.
- */
-async function handleDirectiveNavigation(
-  document: TextDocument,
-  position: { line: number; character: number },
-  uri: string,
-  services: Services,
-  cached: DocumentCacheEntry,
-  log: Logger
-): Promise<Location | null> {
-  const lineText = document
-    .getText({
-      start: { line: position.line, character: 0 },
-      end: { line: position.line + 1, character: 0 },
-    })
-    .trim();
-
-  const filePath = uriToFilePath(uri);
-
-  // Handle #include directives
-  const includeMatch = lineText.match(/^#include\s+["<]([^">]+)[">]/);
-  if (includeMatch && services.bridge?.bridge) {
-    const includePath = includeMatch[1] ?? '';
-    log.debug('Definition: directive include navigation', { includePath });
-
-    try {
-      const result = await services.bridge.bridge.resolveInclude(includePath, filePath);
-      if (result.exists && result.path) {
-        return {
-          uri: `file://${result.path}`,
-          range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        };
-      }
-    } catch (err) {
-      log.debug('Definition: include resolution failed', {
-        includePath,
-        error: err instanceof Error ? err.message : String(err),
-      });
-    }
-    return null;
-  }
-
-  // Handle import statements
-  const importMatch = lineText.match(/^import\s+([^;]+)/);
-  if (importMatch && services.bridge?.bridge) {
-    const modulePath = (importMatch[1] ?? '').trim();
-    log.debug('Definition: directive import navigation', { modulePath });
-
-    // Try stdlib index first
-    const moduleInfo = await services.stdlibIndex?.getModule(modulePath);
-    if (moduleInfo?.resolvedPath) {
-      const targetPath = moduleInfo.filePath ?? moduleInfo.resolvedPath;
-      return {
-        uri: targetPath.startsWith('file://') ? targetPath : `file://${targetPath}`,
-        range: {
-          start: { line: moduleInfo.line ?? 0, character: 0 },
-          end: { line: moduleInfo.line ?? 0, character: 0 },
-        },
-      };
-    }
-
-    // Try bridge resolve_import
-    try {
-      const result = await services.bridge.bridge.resolveImport('import', modulePath, filePath);
-      if (result.exists === 1 && result.path) {
-        return {
-          uri: `file://${result.path}`,
-          range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        };
-      }
-    } catch (err) {
-      log.debug('Definition: import resolution failed', {
-        modulePath,
-        error: err instanceof Error ? err.message : String(err),
-      });
-    }
-    return null;
-  }
-
-  // Handle inherit statements
-  const inheritMatch = lineText.match(/^inherit\s+([^;:]+)/);
-  if (inheritMatch && services.bridge?.bridge) {
-    const className = (inheritMatch[1] ?? '').trim().replace(/['"]/g, '');
-    log.debug('Definition: directive inherit navigation', { className });
-
-    // Try introspection data first (cached inherits from Pike compiler)
-    if (cached.inherits) {
-      const normalizedName = className.replace(/['"]/g, '');
-      const foundInherit = cached.inherits.find(
-        (h: InheritanceInfo) =>
-          h.source_name === normalizedName ||
-          h.path === normalizedName ||
-          h.label === normalizedName ||
-          (h.source_name && h.source_name.replace(/['"]/g, '') === normalizedName)
-      );
-
-      if (foundInherit?.path) {
-        const targetUri = foundInherit.path.startsWith('file://')
-          ? foundInherit.path
-          : `file://${foundInherit.path}`;
-        return {
-          uri: targetUri,
-          range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        };
-      }
-    }
-
-    // Try stdlib index
-    const moduleInfo = await services.stdlibIndex?.getModule(className);
-    if (moduleInfo?.resolvedPath) {
-      const targetPath = moduleInfo.filePath ?? moduleInfo.resolvedPath;
-      return {
-        uri: targetPath.startsWith('file://') ? targetPath : `file://${targetPath}`,
-        range: {
-          start: { line: moduleInfo.line ?? 0, character: 0 },
-          end: { line: moduleInfo.line ?? 0, character: 0 },
-        },
-      };
-    }
-
-    // Try bridge resolve_import for inherit
-    try {
-      const result = await services.bridge.bridge.resolveImport('inherit', className, filePath);
-      if (result.exists === 1 && result.path) {
-        return {
-          uri: `file://${result.path}`,
-          range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        };
-      }
-    } catch (err) {
-      log.debug('Definition: inherit resolution failed', {
-        className,
-        error: err instanceof Error ? err.message : String(err),
-      });
-    }
-
-    // Try with import-qualified names
-    const allImports = cached.symbols?.filter((s: PikeSymbol) => s.kind === 'import') || [];
-    for (const importSymbol of allImports) {
-      const importPath = importSymbol.classname || importSymbol.name;
-      if (importPath) {
-        const qualifiedPath = `${importPath}.${className}`;
-        const qualifiedInfo = await services.stdlibIndex?.getModule(qualifiedPath);
-        if (qualifiedInfo?.resolvedPath) {
-          const targetPath = qualifiedInfo.filePath ?? qualifiedInfo.resolvedPath;
-          return {
-            uri: targetPath.startsWith('file://') ? targetPath : `file://${targetPath}`,
-            range: {
-              start: { line: qualifiedInfo.line ?? 0, character: 0 },
-              end: { line: qualifiedInfo.line ?? 0, character: 0 },
-            },
-          };
-        }
-      }
-    }
-
-    return null;
-  }
-
-  // Handle #require directives
-  const requireMatch = lineText.match(/^#require\s+["<]?([^">;]+)/);
-  if (requireMatch && services.bridge?.bridge) {
-    const requirePath = (requireMatch[1] ?? '').trim();
-    log.debug('Definition: directive require navigation', { requirePath });
-
-    try {
-      const result = await services.bridge.bridge.resolveImport('require', requirePath, filePath);
-      if (result.exists === 1 && result.path) {
-        return {
-          uri: `file://${result.path}`,
-          range: { start: { line: 0, character: 0 }, end: { line: 0, character: 0 } },
-        };
-      }
-    } catch (err) {
-      log.debug('Definition: require resolution failed', {
-        requirePath,
-        error: err instanceof Error ? err.message : String(err),
-      });
-    }
-    return null;
-  }
-
-  // Not a directive line
-  return null;
-}
+// handleDirectiveNavigation moved to definition-directives.ts
 
 /**
  * Find symbol at given position in document.
